refactor(ecs): split up gigantic ecs.go
