+++ internal/pkg/aws/ecs/service.go
@@ -1,0 +1,73 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+// Package ecs provides a client to make API requests to Amazon Elastic Container Service.
+package ecs
+
+import (
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/arn"
+	"github.com/aws/aws-sdk-go/service/ecs"
+)
+
+// Service wraps up ECS Service struct.
+type Service ecs.Service
+
+// ServiceStatus contains the status info of a service.
+type ServiceStatus struct {
+	DesiredCount     int64     `json:"desiredCount"`
+	RunningCount     int64     `json:"runningCount"`
+	Status           string    `json:"status"`
+	LastDeploymentAt time.Time `json:"lastDeploymentAt"`
+	TaskDefinition   string    `json:"taskDefinition"`
+}
+
+// ServiceStatus returns the status of the running service.
+func (s *Service) ServiceStatus() ServiceStatus {
+	return ServiceStatus{
+		Status:           aws.StringValue(s.Status),
+		DesiredCount:     aws.Int64Value(s.DesiredCount),
+		RunningCount:     aws.Int64Value(s.RunningCount),
+		LastDeploymentAt: *s.Deployments[0].UpdatedAt, // FIXME Service assumed to have at least one deployment
+		TaskDefinition:   aws.StringValue(s.Deployments[0].TaskDefinition),
+	}
+}
+
+// ServiceArn is the arn of an ECS service.
+type ServiceArn string
+
+// ClusterName returns the cluster name.
+// For example: arn:aws:ecs:us-west-2:1234567890:service/my-project-test-Cluster-9F7Y0RLP60R7/my-project-test-myService-JSOH5GYBFAIB
+// will return my-project-test-Cluster-9F7Y0RLP60R7
+func (s *ServiceArn) ClusterName() (string, error) {
+	serviceArn := string(*s)
+	parsedArn, err := arn.Parse(serviceArn)
+	if err != nil {
+		return "", err
+	}
+	resources := strings.Split(parsedArn.Resource, "/")
+	if len(resources) != 3 {
+		return "", fmt.Errorf("cannot parse resource for ARN %s", serviceArn)
+	}
+	return resources[1], nil
+}
+
+// ServiceName returns the service name.
+// For example: arn:aws:ecs:us-west-2:1234567890:service/my-project-test-Cluster-9F7Y0RLP60R7/my-project-test-myService-JSOH5GYBFAIB
+// will return my-project-test-myService-JSOH5GYBFAIB
+func (s *ServiceArn) ServiceName() (string, error) {
+	serviceArn := string(*s)
+	parsedArn, err := arn.Parse(serviceArn)
+	if err != nil {
+		return "", err
+	}
+	resources := strings.Split(parsedArn.Resource, "/")
+	if len(resources) != 3 {
+		return "", fmt.Errorf("cannot parse resource for ARN %s", serviceArn)
+	}
+	return resources[2], nil
+}
